Move `Span` and `Event` out of `core`
