+++ tokio-trace/src/span.rs
@@ -196,14 +196,490 @@
 //! [`Attributes`]: ::span::Attributes
 //! [shared span]: ::span::Shared
 //! [`IntoShared`]: ::span::IntoShared
-pub use tokio_trace_core::span::{Attributes, Event, Id, Span, SpanAttributes};
+pub use tokio_trace_core::span::{Attributes, Id, SpanAttributes};
 
 #[cfg(any(test, feature = "test-support"))]
 pub use tokio_trace_core::span::{mock, MockSpan};
 
-use field;
-use std::{fmt, sync::Arc};
-use tokio_trace_core::span::Enter;
+use std::{
+    borrow::Borrow,
+    cmp, fmt,
+    hash::{Hash, Hasher},
+    sync::atomic::{AtomicBool, AtomicUsize, Ordering},
+};
+use {
+    callsite::Callsite,
+    field,
+    subscriber::{Interest, Subscriber},
+    Dispatch, Meta,
+};
+
+/// A handle representing a span, with the capability to enter the span if it
+/// exists.
+///
+/// If the span was rejected by the current `Subscriber`'s filter, entering the
+/// span will silently do nothing. Thus, the handle can be used in the same
+/// manner regardless of whether or not the trace is currently being collected.
+#[derive(PartialEq, Hash)]
+pub struct Span {
+    /// A handle used to enter the span when it is not executing.
+    ///
+    /// If this is `None`, then the span has either closed or was never enabled.
+    inner: Option<Enter>,
+
+    /// Set to `true` when the span closes.
+    ///
+    /// This allows us to distinguish if `inner` is `None` because the span was
+    /// never enabled (and thus the inner state was never created), or if the
+    /// previously entered, but it is now closed.
+    is_closed: bool,
+}
+
+/// `Event`s represent single points in time where something occurred during the
+/// execution of a program.
+///
+/// An event can be compared to a log record in unstructured logging, but with
+/// two key differences:
+/// - Events exist _within the context of a [`Span`]_. Unlike log lines, they may
+///   be located within the trace tree, allowing visibility into the context in
+///   which the event occurred.
+/// - Events have structured key-value data known as _fields_, as well as a
+///   textual message. In general, a majority of the data associated with an
+///   event should be in the event's fields rather than in the textual message,
+///   as the fields are more structed.
+///
+/// [`Span`]: ::span::Span
+#[derive(PartialEq, Hash)]
+pub struct Event<'a> {
+    /// A handle used to enter the span when it is not executing.
+    ///
+    /// If this is `None`, then the span has either closed or was never enabled.
+    inner: Option<Inner<'a>>,
+}
+
+
+/// A handle representing the capacity to enter a span which is known to exist.
+///
+/// Unlike `Span`, this type is only constructed for spans which _have_ been
+/// enabled by the current filter. This type is primarily used for implementing
+/// span handles; users should typically not need to interact with it directly.
+#[derive(Debug)]
+pub struct Inner<'a> {
+    /// The span's ID, as provided by `subscriber`.
+    id: Id,
+
+    /// The subscriber that will receive events relating to this span.
+    ///
+    /// This should be the same subscriber that provided this span with its
+    /// `id`.
+    subscriber: Dispatch,
+
+    /// The span ID of this span's parent, if there is one.
+    parent: Option<Id>,
+
+    /// A flag indicating that the span has been instructed to close when
+    /// possible.
+    ///
+    /// This does _not_ indicate that it is _currently_ okay to close the span,
+    /// only that it _should_ close when it is safe to do so.
+    wants_close: AtomicBool,
+
+    /// A flag that is set when the span is entered for the first time.
+    ///
+    /// If this is false, closing the span should do nothing.
+    has_entered: AtomicBool,
+
+    /// Incremented when new handles are created, and decremented when they are
+    /// dropped if this is the current span.
+    handles: AtomicUsize,
+
+    meta: &'a Meta<'a>,
+}
+
+/// When an `Inner` corresponds to a `Span` rather than an `Event`, it can be
+/// used to enter that span.
+type Enter = Inner<'static>;
+
+/// A guard representing a span which has been entered and is currently
+/// executing.
+///
+/// This guard may be used to exit the span, returning an `Enter` to
+/// re-enter it.
+///
+/// This type is primarily used for implementing span handles; users should
+/// typically not need to interact with it directly.
+#[derive(Debug)]
+#[must_use = "once a span has been entered, it should be exited"]
+struct Entered {
+    id: Id,
+    dispatch: Dispatch,
+    prior: Option<Enter>,
+}
+
+// ===== impl Span =====
+
+impl Span {
+    pub(crate) const NONE: Span = Span {
+        inner: None,
+        is_closed: false,
+    };
+
+    /// Constructs a new `Span` originating from the given [`Callsite`].
+    ///
+    /// The new span will be constructed by the currently-active [`Subscriber`],
+    /// with the [current span] as its parent (if one exists).
+    ///
+    /// If the new span is enabled, then the provided function `if_enabled` is
+    /// envoked on it before it is returned. This allows [field values] and/or
+    /// [`follows_from` annotations] to be added to the span, but skips this
+    /// work for spans which are disabled.
+    ///
+    /// [`Callsite`]: ::callsite::Callsite
+    /// [`Subscriber`]: ::subscriber::Subscriber
+    /// [current span]: ::span::Span::current
+    /// [field values]: ::span::Span::record
+    /// [`follows_from` annotations]: ::span::Span::follows_from
+    #[inline]
+    pub fn new<F>(callsite: &'static dyn Callsite, if_enabled: F) -> Span
+    where
+        F: FnOnce(&mut Span),
+    {
+        let interest = callsite.interest();
+        if interest == Interest::NEVER {
+            return Span::new_disabled();
+        }
+        Dispatch::with_current(|dispatch| {
+            let meta = callsite.metadata();
+            if interest == Interest::SOMETIMES && !dispatch.enabled(meta) {
+                return Span {
+                    inner: None,
+                    is_closed: false,
+                };
+            }
+            let parent = Id::current();
+            let attrs = Attributes::new(parent.clone(), meta);
+            let id = dispatch.new_span(attrs);
+            let inner = Some(Enter::new(id, parent, dispatch, meta));
+            let mut span = Self {
+                inner,
+                is_closed: false,
+            };
+            if_enabled(&mut span);
+            span
+        })
+    }
+
+    /// Constructs a new disabled span.
+    pub fn new_disabled() -> Span {
+        Span {
+            inner: None,
+            is_closed: false,
+        }
+    }
+
+    /// Returns a reference to the span that this thread is currently
+    /// executing.
+    pub fn current() -> Self {
+        Dispatch::with_current(|dispatch| dispatch.current_span().clone())
+    }
+
+    /// Executes the given function in the context of this span.
+    ///
+    /// If this span is enabled, then this function enters the span, invokes
+    /// and then exits the span. If the span is disabled, `f` will still be
+    /// invoked, but in the context of the currently-executing span (if there is
+    /// one).
+    ///
+    /// Returns the result of evaluating `f`.
+    pub fn enter<F: FnOnce() -> T, T>(&mut self, f: F) -> T {
+        match self.inner.take() {
+            Some(inner) => inner.subscriber.as_default(|| {
+                if !self.is_closed() {
+                    inner.take_close();
+                }
+                let guard = inner.enter();
+                let result = f();
+                self.inner = guard.exit();
+                result
+            }),
+            None => f(),
+        }
+    }
+
+    /// Returns the `Id` of the parent of this span, if one exists.
+    pub fn parent(&self) -> Option<Id> {
+        self.inner.as_ref().and_then(Enter::parent)
+    }
+
+    /// Returns a [`Key`](::field::Key) for the field with the given `name`, if
+    /// one exists,
+    pub fn key_for<Q>(&self, name: &Q) -> Option<field::Key<'static>>
+    where
+        Q: Borrow<str>,
+    {
+        self.inner
+            .as_ref()
+            .and_then(|inner| inner.meta.key_for(name))
+    }
+    /// Record a signed 64-bit integer value.
+    pub fn record_value_i64(&mut self, field: &field::Key, value: i64) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_i64(field, value);
+        }
+        self
+    }
+
+    /// Record an umsigned 64-bit integer value.
+    pub fn record_value_u64(&self, field: &field::Key, value: u64) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_u64(field, value);
+        }
+        self
+    }
+
+    /// Record a boolean value.
+    pub fn record_value_bool(&self, field: &field::Key, value: bool) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_bool(field, value);
+        }
+        self
+    }
+
+    /// Record a string value.
+    pub fn record_value_str(&self, field: &field::Key, value: &str) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_str(field, value);
+        }
+        self
+    }
+
+    /// Record a precompiled set of format arguments.
+    pub fn record_value_fmt(&self, field: &field::Key, value: fmt::Arguments) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_fmt(field, value);
+        }
+        self
+    }
+
+    /// Signals that this span should close the next time it is exited, or when
+    /// it is dropped.
+    pub fn close(&mut self) {
+        self.is_closed = true;
+        // Take the span's `Enter` handle and try to close it. If the span is
+        // idle, then the `Enter` will close it immediately when it is dropped.
+        // Otherwise, if corresponds to the same span as the currently-executing
+        // one, it will tell the current span to try to close when it exits.
+        // Taking the `Enter` will prevent this handle from re-entering the span.
+        self.inner.take().as_ref().map(Enter::close);
+    }
+
+    /// Returns `true` if this span is closed.
+    pub fn is_closed(&self) -> bool {
+        self.inner.is_none() && self.is_closed
+    }
+
+    /// Returns `true` if this span was disabled by the subscriber and does not
+    /// exist.
+    pub fn is_disabled(&self) -> bool {
+        self.inner.is_none() && !self.is_closed
+    }
+
+    /// Indicates that the span with the given ID has an indirect causal
+    /// relationship with this span.
+    ///
+    /// This relationship differs somewhat from the parent-child relationship: a
+    /// span may have any number of prior spans, rather than a single one; and
+    /// spans are not considered to be executing _inside_ of the spans they
+    /// follow from. This means that a span may close even if subsequent spans
+    /// that follow from it are still open, and time spent inside of a
+    /// subsequent span should not be included in the time its precedents were
+    /// executing. This is used to model causal relationships such as when a
+    /// single future spawns several related background tasks, et cetera.
+    ///
+    /// If this span is disabled, or the resulting follows-from relationship
+    /// would be invalid, this function will do nothing.
+    pub fn follows_from(&self, from: Id) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.follows_from(from);
+        }
+        self
+    }
+
+    /// Returns this span's `Id`, if it is enabled.
+    pub fn id(&self) -> Option<Id> {
+        self.inner.as_ref().map(Enter::id)
+    }
+
+    /// Returns this span's `Meta`, if it is enabled.
+    pub fn metadata(&self) -> Option<&'static Meta<'static>> {
+        self.inner.as_ref().map(|inner| inner.metadata())
+    }
+
+    fn clone(&self) -> Self {
+        let inner = self.inner.as_ref().map(Enter::duplicate_handle);
+        Self {
+            inner,
+            is_closed: self.is_closed,
+        }
+    }
+}
+
+impl fmt::Debug for Span {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let mut span = f.debug_struct("Span");
+        if let Some(ref inner) = self.inner {
+            span.field("id", &inner.id())
+                .field("parent", &inner.parent())
+        } else {
+            span.field("disabled", &true)
+        }.finish()
+    }
+}
+
+// ===== impl Event =====
+impl<'a> Event<'a> {
+    /// Constructs a new `Span` originating from the given [`Callsite`].
+    ///
+    /// The new span will be constructed by the currently-active [`Subscriber`],
+    /// with the [current span] as its parent (if one exists).
+    ///
+    /// If the new span is enabled, then the provided function `if_enabled` is
+    /// envoked on it before it is returned. This allows [field values] and/or
+    /// [`follows_from` annotations] to be added to the span, but skips this
+    /// work for spans which are disabled.
+    ///
+    /// [`Callsite`]: ::callsite::Callsite
+    /// [`Subscriber`]: ::subscriber::Subscriber
+    /// [current span]: ::span::Span::current
+    /// [field values]: ::span::Span::record
+    /// [`follows_from` annotations]: ::span::Span::follows_from
+    #[inline]
+    pub fn new<F>(callsite: &'a dyn Callsite, if_enabled: F) -> Self
+    where
+        F: FnOnce(&mut Self),
+    {
+        let interest = callsite.interest();
+        if interest == Interest::NEVER {
+            return Self { inner: None };
+        }
+        Dispatch::with_current(|dispatch| {
+            let meta = callsite.metadata();
+            if interest == Interest::SOMETIMES && !dispatch.enabled(meta) {
+                return Self { inner: None };
+            }
+            let parent = Id::current();
+            let attrs = Attributes::new(parent.clone(), meta);
+            let id = dispatch.new_id(attrs);
+            let inner = Enter::new(id, parent, dispatch, meta);
+            inner.has_entered.store(true, Ordering::Relaxed);
+            inner.close();
+            let mut event = Self { inner: Some(inner) };
+            if_enabled(&mut event);
+            event
+        })
+    }
+
+    /// Adds a formattable message describing the event that occurred.
+    pub fn message(&mut self, key: &field::Key, message: fmt::Arguments) -> &mut Self {
+        if let Some(ref mut inner) = self.inner {
+            inner.subscriber.record_fmt(&inner.id, key, message);
+        }
+        self
+    }
+
+    /// Record a signed 64-bit integer value.
+    pub fn record_value_i64(&mut self, field: &field::Key, value: i64) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_i64(field, value);
+        }
+        self
+    }
+
+    /// Record an umsigned 64-bit integer value.
+    pub fn record_value_u64(&self, field: &field::Key, value: u64) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_u64(field, value);
+        }
+        self
+    }
+
+    /// Record a boolean value.
+    pub fn record_value_bool(&self, field: &field::Key, value: bool) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_bool(field, value);
+        }
+        self
+    }
+
+    /// Record a string value.
+    pub fn record_value_str(&self, field: &field::Key, value: &str) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_str(field, value);
+        }
+        self
+    }
+
+    /// Record a precompiled set of format arguments.
+    pub fn record_value_fmt(&self, field: &field::Key, value: fmt::Arguments) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.record_value_fmt(field, value);
+        }
+        self
+    }
+
+    /// Returns the `Id` of the parent of this span, if one exists.
+    pub fn parent(&self) -> Option<Id> {
+        self.inner.as_ref().and_then(Enter::parent)
+    }
+
+    /// Returns a [`Key`](::field::Key) for the field with the given `name`, if
+    /// one exists,
+    pub fn key_for<Q>(&self, name: &Q) -> Option<field::Key<'a>>
+    where
+        Q: Borrow<str>,
+    {
+        self.inner
+            .as_ref()
+            .and_then(|inner| inner.meta.key_for(name))
+    }
+
+    /// Returns `true` if this span was disabled by the subscriber and does not
+    /// exist.
+    pub fn is_disabled(&self) -> bool {
+        self.inner.is_none()
+    }
+
+    /// Indicates that the span with the given ID has an indirect causal
+    /// relationship with this event.
+    ///
+    /// This relationship differs somewhat from the parent-child relationship: a
+    /// span may have any number of prior spans, rather than a single one; and
+    /// spans are not considered to be executing _inside_ of the spans they
+    /// follow from. This means that a span may close even if subsequent spans
+    /// that follow from it are still open, and time spent inside of a
+    /// subsequent span should not be included in the time its precedents were
+    /// executing. This is used to model causal relationships such as when a
+    /// single future spawns several related background tasks, et cetera.
+    ///
+    /// If this event is disabled, or the resulting follows-from relationship
+    /// would be invalid, this function will do nothing.
+    pub fn follows_from(&self, from: Id) -> &Self {
+        if let Some(ref inner) = self.inner {
+            inner.follows_from(from);
+        }
+        self
+    }
+
+    /// Returns this span's `Id`, if it is enabled.
+    pub fn id(&self) -> Option<Id> {
+        self.inner.as_ref().map(Enter::id)
+    }
+
+    /// Returns this span's `Meta`, if it is enabled.
+    pub fn metadata(&self) -> Option<&'a Meta<'a>> {
+        self.inner.as_ref().map(|inner| inner.metadata())
+    }
+}
 
 /// Trait for converting a `Span` into a cloneable `Shared` span.
 pub trait IntoShared {
@@ -317,6 +793,278 @@ impl Shared {
             inner.follows_from(from);
         }
         self
+    }
+}
+
+
+// ===== impl Enter =====
+
+impl<'a> Inner<'a> {
+    /// Indicates the span _should_ be closed the next time it exits or this
+    /// handle is dropped.
+    pub fn close(&self) {
+        self.wants_close.store(true, Ordering::Release);
+    }
+
+    /// Indicates that the span with the given ID has an indirect causal
+    /// relationship with this span.
+    ///
+    /// This relationship differs somewhat from the parent-child relationship: a
+    /// span may have any number of prior spans, rather than a single one; and
+    /// spans are not considered to be executing _inside_ of the spans they
+    /// follow from. This means that a span may close even if subsequent spans
+    /// that follow from it are still open, and time spent inside of a
+    /// subsequent span should not be included in the time its precedents were
+    /// executing. This is used to model causal relationships such as when a
+    /// single future spawns several related background tasks, et cetera.
+    ///
+    /// If this span is disabled, this function will do nothing. Otherwise, it
+    /// returns `Ok(())` if the other span was added as a precedent of this
+    /// span, or an error if this was not possible.
+    pub fn follows_from(&self, from: Id) {
+        self.subscriber.add_follows_from(&self.id, from)
+    }
+
+    /// Returns the span's ID.
+    pub fn id(&self) -> Id {
+        self.id.clone()
+    }
+
+    /// Returns the ID of the span's parent, if it has one.
+    pub fn parent(&self) -> Option<Id> {
+        self.parent.clone()
+    }
+
+    /// Returns the span's metadata.
+    pub fn metadata(&self) -> &'a Meta<'a> {
+        self.meta
+    }
+
+    /// Record a signed 64-bit integer value.
+    pub(crate) fn record_value_i64(&self, field: &field::Key, value: i64) {
+        if self.meta.contains_key(field) {
+            self.subscriber.record_i64(&self.id, field, value)
+        }
+    }
+
+    /// Record an umsigned 64-bit integer value.
+    pub(crate) fn record_value_u64(&self, field: &field::Key, value: u64) {
+        if self.meta.contains_key(field) {
+            self.subscriber.record_u64(&self.id, field, value)
+        }
+    }
+
+    /// Record a boolean value.
+    pub(crate) fn record_value_bool(&self, field: &field::Key, value: bool) {
+        if self.meta.contains_key(field) {
+            self.subscriber.record_bool(&self.id, field, value)
+        }
+    }
+
+    /// Record a string value.
+    pub(crate) fn record_value_str(&self, field: &field::Key, value: &str) {
+        if self.meta.contains_key(field) {
+            self.subscriber.record_str(&self.id, field, value)
+        }
+    }
+
+    /// Record a precompiled set of format arguments value.
+    pub(crate) fn record_value_fmt(&self, field: &field::Key, value: fmt::Arguments) {
+        if self.meta.contains_key(field) {
+            self.subscriber.record_fmt(&self.id, field, value)
+        }
+    }
+
+    fn new(id: Id, parent: Option<Id>, subscriber: &Dispatch, meta: &'a Meta<'a>) -> Self {
+        Self {
+            id,
+            subscriber: subscriber.clone(),
+            parent,
+            wants_close: AtomicBool::from(false),
+            has_entered: AtomicBool::from(false),
+            handles: AtomicUsize::from(1),
+            meta,
+        }
+    }
+
+    fn duplicate(&self) -> Self {
+        let id = self.subscriber.clone_span(self.id.clone());
+        Self {
+            id,
+            subscriber: self.subscriber.clone(),
+            parent: self.parent.clone(),
+            wants_close: AtomicBool::from(self.wants_close()),
+            has_entered: AtomicBool::from(self.has_entered()),
+            handles: AtomicUsize::from(self.handle_count()),
+            meta: self.meta,
+        }
+    }
+
+    /// Returns `true` if this span _should_ close itself.
+    ///
+    /// This is true IFF:
+    /// - the span has been told to close,
+    /// - the span has been entered already (we don't want to close spans that
+    ///   have never been entered),
+    /// - there aren't multiple handles capable of entering the span.
+    fn should_close(&self) -> bool {
+        self.wants_close() && self.has_entered() && self.handle_count() == 1
+    }
+
+    fn has_entered(&self) -> bool {
+        self.has_entered.load(Ordering::Acquire)
+    }
+
+    fn wants_close(&self) -> bool {
+        self.wants_close.load(Ordering::Acquire)
+    }
+
+    fn take_close(&self) -> bool {
+        self.wants_close.swap(false, Ordering::Release)
+    }
+
+    fn handle_count(&self) -> usize {
+        self.handles.load(Ordering::Acquire)
+    }
+}
+
+impl Enter {
+    /// Consumes `span` and returns the inner entering handle, if the span is
+    /// enabled.
+    ///
+    /// The returned `Enter` has approximately the same behaviour to a `Span`.
+    /// However, it is primarily intended for use in libraries custom span
+    /// types; the `Span` handle will typically represent a more ergonomic API
+    /// for actually _using_ spans.
+    pub fn from_span(span: Span) -> Option<Self> {
+        span.inner
+    }
+
+    /// Return a new `Span` handle to the span represented by this `Enter`.
+    pub fn into_span(self) -> Span {
+        Span {
+            inner: Some(self),
+            is_closed: false,
+        }
+    }
+
+    /// Enters the span, returning a guard that may be used to exit the span and
+    /// re-enter the prior span.
+    ///
+    /// This is used internally to implement `Span::enter`. It may be used for
+    /// writing custom span handles, but should generally not be called directly
+    /// when entering a span.
+    pub fn enter(&self) -> Entered {
+        // The current handle will no longer enter the span, since it has just
+        // been used to enter. Therefore, it will be safe to close the span if
+        // no additional handles exist when the span is exited.
+        self.handles.fetch_sub(1, Ordering::Release);
+        // The span has now been entered, so it's okay to close it.
+        self.has_entered.store(true, Ordering::Release);
+        let Span { inner: prior, .. } = self.subscriber.enter(self.duplicate().into_span());
+        Entered {
+            prior,
+            id: self.id.clone(),
+            dispatch: self.subscriber.clone(),
+        }
+    }
+
+    /// Exits the span entry represented by an `Entered` guard, consuming it,
+    /// and updates `self` to canonically represent the referenced span's state
+    /// after the other entry has exited.
+    ///
+    /// If the other handle to the span wanted to close the span on exit, it
+    /// will not do so. Instead, the responsibility for performing the `close`
+    /// will be transferred to `self` --- if `self` did not previously want to
+    /// close, but `other` did, `self` will now want to close.
+    ///
+    /// This means that dropping `other` will no longer close the span, even if
+    /// it previously would have.
+    ///
+    /// This function is intended to be used by span handle implementations to
+    /// ensuring that multiple entering handles are kept consistent. Probably
+    /// don't use this unless you know what you're doing.
+    pub fn exit_and_join(&self, other: Entered) {
+        if let Some(other) = other.exit() {
+            self.handles.store(other.handle_count(), Ordering::Release);
+            self.has_entered
+                .store(other.has_entered(), Ordering::Release);
+            self.wants_close
+                .store(other.take_close(), Ordering::Release);
+        }
+    }
+
+    fn duplicate_handle(&self) -> Self {
+        self.handles.fetch_add(1, Ordering::Release);
+        self.duplicate()
+    }
+}
+
+impl<'a> cmp::PartialEq for Inner<'a> {
+    fn eq(&self, other: &Self) -> bool {
+        self.id == other.id
+    }
+}
+
+impl<'a> Hash for Inner<'a> {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.id.hash(state);
+    }
+}
+
+impl<'a> Drop for Inner<'a> {
+    fn drop(&mut self) {
+        self.subscriber.drop_span(self.id.clone());
+        // If this handle wants to be closed, try to close it --- either by
+        // closing it now if it is idle, or telling the current span to close
+        // when it exits, if it is the current span.
+        if self.has_entered() && self.wants_close() {
+            match self.subscriber.current_span().inner {
+                // If the `enter` being dropped corresponds to the same span as
+                // the current span, then we cannot close it yet. Instead, we
+                // signal to the copy of this span that currently occupies
+                // CURRENT_SPAN that it should try to close when it is exited.
+                Some(ref current) if current == self => {
+                    current.handles.fetch_sub(1, Ordering::Release);
+                    current
+                        .wants_close
+                        .store(self.wants_close(), Ordering::Release);
+                }
+                // If this span is not the current span, then it may close now,
+                // if there are no other handles with the capacity to re-enter it.
+                _ if self.handle_count() <= 1 => {
+                    self.subscriber.close(self.id());
+                }
+                _ => {}
+            }
+        }
+    }
+}
+
+impl Entered {
+    /// Exit the `Entered` guard, returning an `Enter` handle that may be used
+    /// to re-enter the span, or `None` if the span closed while performing the
+    /// exit.
+    pub fn exit(self) -> Option<Enter> {
+        let prior = self
+            .prior
+            .map(Enter::into_span)
+            .unwrap_or_else(Span::new_disabled);
+        Enter::from_span(self.dispatch.exit(self.id, prior)).and_then(|inner| {
+            if inner.should_close() {
+                // Dropping `inner` will allow it to perform the closure if
+                // able.
+                None
+            } else {
+                // We are returning a new `Enter`. Increment the number of
+                // handles that may enter the span.
+                inner.handles.fetch_add(1, Ordering::Release);
+                // The span will want to close if it is dropped prior to
+                // being re-entered.
+                inner.wants_close.store(true, Ordering::Release);
+                Some(inner)
+            }
+        })
     }
 }
